+++ crates/apollo-compiler/src/execution/introspection_split.rs
@@ -65,6 +65,7 @@ pub enum SchemaIntrospectionSplit {
 #[derive(Debug)]
 pub enum SchemaIntrospectionError {
     SuspectedValidationBug(SuspectedValidationBug),
+    DeeplyNestedIntrospectionList(Option<SourceSpan>),
     Unsupported {
         message: String,
         location: Option<SourceSpan>,
@@ -123,7 +124,9 @@ impl SchemaIntrospectionSplit {
                 .collect();
             let introspection_document =
                 make_single_operation_document(schema, document, new_operation, fragments);
-            Ok(Self::Only(SchemaIntrospectionQuery(introspection_document)))
+            Ok(Self::Only(
+                SchemaIntrospectionQuery::assume_only_intropsection_fields(introspection_document)?,
+            ))
         } else {
             let mut fragments_done = HashSet::with_hasher(Default::default());
             let mut new_documents = Split {
@@ -136,11 +139,13 @@ impl SchemaIntrospectionSplit {
                 &operation.selection_set,
             );
             Ok(Self::Both {
-                introspection_query: SchemaIntrospectionQuery(new_documents.introspection.build(
-                    schema,
-                    document,
-                    operation_selection_set.introspection,
-                )),
+                introspection_query: SchemaIntrospectionQuery::assume_only_intropsection_fields(
+                    new_documents.introspection.build(
+                        schema,
+                        document,
+                        operation_selection_set.introspection,
+                    ),
+                )?,
                 filtered_document: new_documents.other.build(
                     schema,
                     document,
@@ -244,7 +249,7 @@ fn make_single_operation_document(
         .expect("filtering a valid document should result in a valid document")
 }
 
-fn get_fragment<'doc>(
+pub(crate) fn get_fragment<'doc>(
     document: &'doc Valid<ExecutableDocument>,
     name: &Name,
 ) -> Result<&'doc Node<Fragment>, SchemaIntrospectionError> {
@@ -592,6 +597,9 @@ impl SchemaIntrospectionError {
     pub fn into_graphql_error(self, sources: &SourceMap) -> GraphQLError {
         match self {
             Self::SuspectedValidationBug(s) => s.into_graphql_error(sources),
+            Self::DeeplyNestedIntrospectionList(location) => {
+                GraphQLError::new("Maximum introspection depth exceeded", location, sources)
+            }
             Self::Unsupported { message, location } => {
                 GraphQLError::new(message, location, sources)
             }